app/backend/requirements.txt
langchain
langchain-openai
langchain_pinecone
langchain[docarray]
docarray
pydantic
pytube 
python-dotenv
tiktoken 
pinecone-client 
scikit-learn
ruff
pypdf